Added a sentry and chasing bullets.

+++ core/src/ga/gussio/ld38/earthinvaders/entities/Bullet.java
@@ -1,5 +1,6 @@
 package ga.gussio.ld38.earthinvaders.entities;
 
+import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
@@ -11,7 +12,7 @@ public class Bullet extends Entity {
     protected int speed = 8;
     protected int lifetime = 10000;
 
-    public Bullet(int x, int y, double angle) {
+    public Bullet(float x, float y, double angle) {
         super(x, y);
 /*		velX = (int) (speed*Math.cos(angle+Math.toRadians(90)));
 //		velY = (int) (speed*Math.sin(angle+Math.toRadians(90)));*/
@@ -50,14 +51,13 @@ public class Bullet extends Entity {
     }
 
     public void checkCollision(){
-        for(Entity e : GameScreen.entities){
-            if(e instanceof Meteorite){
+        for(Entity e : GameScreen.entities) {
+            if (e instanceof Meteorite) {
                 Meteorite m = (Meteorite) e;
-                double dx = Math.abs(x-m.collision.getXCenter());
-                double dy = Math.abs(y-m.collision.getYCenter());
+                double dx = Math.abs(x - m.collision.getXCenter());
+                double dy = Math.abs(y - m.collision.getYCenter());
                 double radius = m.collision.getRadius();
-
-                if(Math.pow(dx, 2) + Math.pow(dy, 2) <= Math.pow(radius, 2) || dx + dy <= radius){
+                if (Math.pow(dx, 2) + Math.pow(dy, 2) <= Math.pow(radius, 2) || dx + dy <= radius) {
                     m.damage();
                     GameScreen.addScore(5);
                     GameScreen.entities.remove(this);
